If we look up a patient in PDS and they're now marked as valid (and
previously were invalid), we should add them to any upcoming sessions if
approriate. This uses a `SchoolMove` instance to handle this to apply
the same logic used elsewhere in the service.

+++ app/models/school_move.rb
@@ -53,7 +53,7 @@ class SchoolMove < ApplicationRecord
     ActiveRecord::Base.transaction do
       update_patient!
       update_sessions!(move_to_school:)
-      SchoolMove.where(patient:).destroy_all
+      SchoolMove.where(patient:).destroy_all if persisted?
     end
   end
